+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.1.0.0
 
   * update all function names to be more idiomatic
+  * add cmdlet binding to all functions, output will use write-verbose
 
 ## 0.0.3.10